% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogf.R
\name{get_gf_python}
\alias{get_gf_python}
\title{use import_from_python in reticulate to get the geneformer modules}
\usage{
get_gf_python()
}
\description{
use import_from_python in reticulate to get the geneformer modules
}
