% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogf.R
\name{get_gf_safetensors}
\alias{get_gf_safetensors}
\title{retrieve or place the model.safetensors for Geneformer (6L?) in cache, zipped}
\usage{
get_gf_safetensors(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of BiocFileCache::BiocFileCache()}
}
\value{
character path to zip file once cached
}
\description{
retrieve or place the model.safetensors for Geneformer (6L?) in cache, zipped
}
\note{
A README.txt is included in the zip file
}
\examples{
pa = get_gf_safetensors()
td = tempdir()
unzip(pa, exdir=td)
nwts_safet(file.path(td, "GF_safetensors", "model.safetensors"))
st = get_safet() # not standard, hands live python ref
oo = st$safe_open(file.path(td, "GF_safetensors", "model.safetensors"), "pt")
mk = oo$keys()
length(mk)
head(sapply(mk, function(z) oo$get_tensor(z)$shape))
}
