% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogf.R
\name{get_panc_loom_path}
\alias{get_panc_loom_path}
\title{retrieve or place a loom file for Grun/Muraro's pancreas study in BiocFileCache}
\usage{
get_panc_loom_path(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of BiocFileCache::BiocFileCache()}
}
\value{
character path to loom file
}
\description{
retrieve or place a loom file for Grun/Muraro's pancreas study in BiocFileCache
}
\note{
A loom file was distributed at
\url{https://figshare.com/articles/dataset/Data_used_for_demo_of_the_code_accompanying_the_i_Assessing_the_limits_of_zero-shot_foundation_models_in_single-cell_biology_i_paper_/24747228} under
data/datasets/geneformer/pancreas_scib.  The relevant publication
seems to be
 \url{https://pubmed.ncbi.nlm.nih.gov/27345837/}, GSE81076, etc.
}
\examples{
pa = get_panc_loom_path()
proc = basilisk::basiliskStart(BiocGF:::gfenv)
x = basilisk::basiliskRun(proc, function() {
 anndata::read_loom(pa)
})
basilisk::basiliskStop(proc)
x
}
