% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogf.R
\name{get_panc_tok_path}
\alias{get_panc_tok_path}
\title{retrieve or place a zipfile with tokenization dataset for Grun/Muraro's pancreas study in BiocFileCache}
\usage{
get_panc_tok_path(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{cache instance likely inheriting from BiocFileCache::BiocFileCache}
}
\value{
character path to zip file
}
\description{
retrieve or place a zipfile with tokenization dataset for Grun/Muraro's pancreas study in BiocFileCache
}
\note{
A loom file was distributed at
\url{https://figshare.com/articles/dataset/Data_used_for_demo_of_the_code_accompanying_the_i_Assessing_the_limits_of_zero-shot_foundation_models_in_single-cell_biology_i_paper_/24747228} under
data/datasets/geneformer/pancreas_scib.  The relevant publication
seems to be
 \url{https://pubmed.ncbi.nlm.nih.gov/27345837/}, GSE81076, etc.
}
