% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogf.R
\name{getgf}
\alias{getgf}
\title{tickle basilisk installation}
\usage{
getgf()
}
\description{
tickle basilisk installation
}
\note{
Intent is to hand back a python reference, not endorsed in basilisk.
Use better insulation in production.
}
\examples{
if (interactive()) {
 cur = getwd()
 td = tempdir()
 setwd(td)
 pa = get_panc_tok_path()
 unzip(pa)  # destination = "."
 ds = getgf() # get python module
 ref = ds$load_from_disk("panc_hfds/tbyr.dataset")
 print(ref)
 setwd(cur)
}
}
