% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_tokenize_loom.R
\name{gf_tokenize_h5ad}
\alias{gf_tokenize_h5ad}
\title{tokenize a h5ad file}
\usage{
gf_tokenize_h5ad(
  h5adpath,
  output_directory = tempdir(),
  output_prefix = "h5adtoks"
)
}
\arguments{
\item{h5adpath}{character(1) path to h5ad file}

\item{output_directory}{character(1) defaults to tempdir()}

\item{output_prefix}{character(1) defaults to "h5adtoks"}
}
\description{
tokenize a h5ad file
}
\note{
The python functionality calls for dirname(h5adpath).
}
