% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogf.R
\name{gf_tokenize_loom}
\alias{gf_tokenize_loom}
\title{tokenize a loom file}
\usage{
gf_tokenize_loom(
  loompath,
  output_directory = tempdir(),
  output_prefix = "loomtoks"
)
}
\arguments{
\item{loompath}{character(1) path to loom file}

\item{output_directory}{character(1) defaults to tempdir()}

\item{output_prefix}{character(1) defaults to "loomtoks"}
}
\description{
tokenize a loom file
}
\note{
The python functionality calls for dirname(loompath).
}
\examples{
lp = get_panc_loom_path()
gf_tokenize_loom(lp)
dir(tempdir())
}
