% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_safetensmod.R
\name{nwts_safet}
\alias{nwts_safet}
\title{count number of weights in safetensors file}
\usage{
nwts_safet(fn = "model.safetensors")
}
\arguments{
\item{fn}{character(1) filename defaults to model.safetensors}
}
\description{
count number of weights in safetensors file
}
