% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogf.R
\name{panc_tok_zip_url}
\alias{panc_tok_zip_url}
\title{give URL for 'dataset' corresponding to tokenization of pancreas data}
\usage{
panc_tok_zip_url()
}
\description{
give URL for 'dataset' corresponding to tokenization of pancreas data
}
