% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usebert.R
\name{usebert}
\alias{usebert}
\title{work with a template python script to illustrate mask substitution by a BERT model}
\usage{
usebert(phrase = "The capital of Spain is [MASK].", decode = TRUE)
}
\arguments{
\item{phrase}{character(1) a string with `[MASK]` ending with a `.`.}

\item{decode}{logical(1) if TRUE, decoding is conducted; otherwise
a python reference is handed back}
}
\description{
work with a template python script to illustrate mask substitution by a BERT model
}
\examples{
usebert()
}
