% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usebert.R
\name{usebert_pipeline}
\alias{usebert_pipeline}
\title{work with a template python script to illustrate mask substitution by a BERT model,
and use the pipeline unmasker to produce a list of scored substitutions}
\usage{
usebert_pipeline(phrase = "The capital of Spain is [MASK].")
}
\arguments{
\item{phrase}{character(1) a string with `[MASK]` ending with a `.`.}
}
\description{
work with a template python script to illustrate mask substitution by a BERT model,
and use the pipeline unmasker to produce a list of scored substitutions
}
\examples{
usebert_pipeline()
}
